some auth working

+++ backend/routes/events.js
@@ -8,7 +8,8 @@ router.post('', checkAuth, (req, res, next) => {
         title: req.body.title,
         location: req.body.location,
         date: req.body.date,
-        description: req.body.description
+        description: req.body.description,
+        creator: req.userData.userId
     });
     event.save().then(createdEvent => {
         res.status(201).json({
@@ -28,9 +29,13 @@ router.get('', (req, res, next) => {
 })
 
 router.delete('/:id', checkAuth, (req, res, next) => {
-    Event.deleteOne({ _id: req.params.id }).then(() => {
-        console.log("Deleted on server")
-        res.status(200).json({ message: 'Event deleted!' });
+    Event.deleteOne({ _id: req.params.id, creator: req.userData.userId }).then(result => {
+        if (result.deletedCount > 0) {
+          console.log("Deleted on server");
+          res.status(200).json({message: "Deletion successful!"});
+        } else {
+          res.status(401).json({ message: "Not authorized." })
+        }
     })
 });
 
@@ -42,9 +47,12 @@ router.put('/:id', checkAuth, (req, res, next) => {
         date: req.body.date,
         description: req.body.description
     })
-    Event.updateOne({ _id: req.params.id }, event).then(result => {
-        console.log(event);
-        res.status(200).json({ message: "Updated successfully." })
+    Event.updateOne({ _id: req.params.id, creator: req.userData.userId }, event).then(result => {
+      if (result.modifiedCount > 0) {
+        res.status(200).json({message: "Update successful!"});
+      } else {
+        res.status(401).json({ message: "Not authorized." })
+      }
     })
 })
